Integration of Azure App service API in the catbingo Game
I have created the API to access the Azure SQL database and published the API to Azure App Service. I have made scripts in catbingo game to access the API and perform the CRUD operations. I have also made changes to the game to assign the game state fetched from API to the current game, update the state in Azure database, post new user profiles and delete players. Also created the input field to take the name of the new player and created a horizontal scroll view to give the options of selecting the players.